Rename interface-file name
